Pruebas/Prueba_Entrada/app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
click==8.1.8
colorama==0.4.6
databases==0.9.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
pytest==7.4.0
Flask==2.3.3
python-dotenv==1.0.0
bandit==1.7.0
locust==2.17.0